requirements.txt
# Core Dependencies
scapy>=2.5.0
python-nmap>=0.7.1
impacket>=0.11.0
paramiko>=3.4.0
requests>=2.31.0
urllib3>=2.0.0

# Machine Learning (Python 3.12 compatible)
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0

# Reinforcement Learning
gymnasium>=0.29.0

# Cryptography & Security
cryptography>=41.0.0
pycryptodome>=3.19.0

# Networking
netifaces>=0.11.0
psutil>=5.9.0
dnspython>=2.4.0
networkx>=3.0

# Visualization
matplotlib>=3.7.0

# C2 Server
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Utilities
colorama>=0.4.6
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0